maint: add empty line after package statement to accomodate $VERSION
PkgVersion will add a new line after package statement with version
number.  If an empty line exists, it uses that line otherwise it creates
a new line which means that error messages from releases will not match
development sources.  This change adds that empty line.

Note that the change is kind of pointless because of other distzilla
plugins that also add lines at the top of the file (only authority at the
moment but maybe others in the future) have the same issue but do make use
of an empty line if available.  However, this change removes the warning
coming from PkgVersion (see issue rjbs/Dist-Zilla#272).